feat: show battery with decimal precision

+++ heltec-controller-receiver/include/battery.h
@@ -8,7 +8,7 @@ class Battery {
 public:
     Battery();
     void setup();
-    int getPercentage();
+    float getPercentage();
     float getVoltage();
     bool isCharging();
